requirements.txt
playwright 
beautifulsoup4 
requests 
supabase 
pdfminer.six 
pytesseract 
opencv-python
python-dotenv
groq
openai>=1.0.0
tenacity

# Developer Dependencies 
pytest 
black 
isort 

# pdf
pdfplumber 
pytesseract 
Pillow
